refactor(Lotto, Number): model 패키지로 이동
