+++ src/test/java/step3/LottoTest.java
@@ -2,6 +2,8 @@ package step3;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step3.model.Lotto;
+import step3.model.Number;
 
 import java.util.ArrayList;
 import java.util.List;
